refactor: collapse extension packages into root project
