+++ .vscode/tasks.json
@@ -26,7 +26,7 @@
     {
       "label": "npm: watch",
       "type": "shell",
-      "command": "npm --prefix ${workspaceFolder}/../../ run watch",
+      "command": "npm run watch",
       "isBackground": true,
       "group": {
         "kind": "build",
