+++ .vscodeignore
@@ -1,6 +1,6 @@
 **/*
 !package.json
 !README.md
-!LICENSE.txt
+!LICENSE
 !textlint-icon_128x128.png
 !dist/*.js
