+++ docs/README_EXTENSION.md
@@ -1,1 +1,1 @@
-../packages/textlint/README.md
+../README.md
