+++ package.json
@@ -1,40 +1,179 @@
 {
-  "name": "vscode-textlint-parent",
-  "private": true,
-  "workspaces": [
-    "packages/textlint",
-    "packages/textlint-server"
+  "name": "textlint",
+  "displayName": "textlint",
+  "version": "0.12.1",
+  "description": "Integrates textlint into VS Code.",
+  "categories": [
+    "Linters"
   ],
+  "homepage": "https://github.com/textlint/vscode-textlint",
+  "bugs": {
+    "url": "https://github.com/textlint/vscode-textlint/issues"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/textlint/vscode-textlint"
+  },
+  "license": "MIT",
+  "publisher": "3w36zj6",
+  "main": "./dist/extension",
   "scripts": {
-    "clean": "npm run clean --ws",
-    "tsc": "npm run tsc --ws",
-    "type-check": "npm exec --ws -- tsc --noEmit",
+    "clean": "rimraf dist out",
+    "tsc": "tsc -p ./",
+    "type-check": "tsc --noEmit",
     "build": "webpack --mode production",
     "watch": "webpack --mode development --watch",
-    "test": "npm run test --ws --if-present",
-    "format": "npm exec --ws -- prettier --check .  --config ../../package.json --ignore-path ../../.gitignore",
-    "format:write": "npm exec --ws -- prettier --write .  --config ../../package.json --ignore-path ../../.gitignore",
-    "lint": "npm exec --ws -- eslint --config .eslintrc.json '**/*.ts'",
-    "lint:fix": "npm exec --ws -- eslint --config .eslintrc.json --fix '**/*.ts'",
+    "pretest": "npm run tsc",
+    "test": "node ./out/test/runTest.js",
+    "format": "prettier --check . --config package.json --ignore-path .gitignore",
+    "format:write": "prettier --write . --config package.json --ignore-path .gitignore",
+    "lint": "eslint --config .eslintrc.json '**/*.ts'",
+    "lint:fix": "eslint --config .eslintrc.json --fix '**/*.ts'",
     "fix": "run-s format:write lint:fix",
-    "sort": "npm exec --include-workspace-root --ws -- sort-package-json",
-    "version": "npm version --ws ",
-    "upgrade": "npm exec --ws -- ncu -u",
+    "sort": "sort-package-json",
+    "upgrade": "ncu -u",
     "prepackage": "npm run build",
-    "package": "npm run package -w packages/textlint"
+    "package": "vsce package --no-dependencies"
+  },
+  "contributes": {
+    "commands": [
+      {
+        "title": "Fix all auto-fixable Problems",
+        "category": "textlint",
+        "command": "textlint.executeAutofix"
+      },
+      {
+        "title": "Create '.textlintrc' File",
+        "category": "textlint",
+        "command": "textlint.createConfig"
+      },
+      {
+        "title": "Show Output Channel",
+        "category": "textlint",
+        "command": "textlint.showOutputChannel"
+      }
+    ],
+    "configurationDefaults": {
+      "files.associations": {
+        ".textlintrc.json": "jsonc",
+        ".textlintignore": "ignore"
+      }
+    },
+    "configuration": {
+      "type": "object",
+      "title": "textlint",
+      "properties": {
+        "textlint.languages": {
+          "default": [
+            "markdown",
+            "plaintext",
+            "html",
+            "tex",
+            "latex",
+            "doctex"
+          ],
+          "type": [
+            "array"
+          ],
+          "items": {
+            "type": "string"
+          },
+          "description": "Languages to lint with textlint."
+        },
+        "textlint.configPath": {
+          "type": "string",
+          "default": null,
+          "description": "An absolute path to textlint config file."
+        },
+        "textlint.ignorePath": {
+          "type": "string",
+          "default": null,
+          "description": "An absolute path to textlint ignore file."
+        },
+        "textlint.nodePath": {
+          "type": "string",
+          "default": null,
+          "description": "A path added to NODE_PATH when resolving the textlint module."
+        },
+        "textlint.run": {
+          "type": "string",
+          "enum": [
+            "onSave",
+            "onType"
+          ],
+          "default": "onSave",
+          "description": "Run the linter on save (onSave) or on type (onType)"
+        },
+        "textlint.autoFixOnSave": {
+          "type": "boolean",
+          "default": false,
+          "description": "Turns auto fix on save on or off."
+        },
+        "textlint.trace": {
+          "type": "string",
+          "enum": [
+            "off",
+            "messages",
+            "verbose"
+          ],
+          "default": "off",
+          "description": "Traces the communication between VSCode and the textlint linter service."
+        },
+        "textlint.targetPath": {
+          "type": "string",
+          "default": "",
+          "description": "Target files path that runs lint."
+        }
+      }
+    }
+  },
+  "activationEvents": [
+    "onLanguage",
+    "onCommand:textlint.showOutputChannel",
+    "onCommand:textlint.createConfig",
+    "onCommand:textlint.executeAutofix"
+  ],
+  "dependencies": {
+    "glob": "^7.2.0",
+    "minimatch": "^3.0.4",
+    "vscode-languageclient": "^7.0.0",
+    "vscode-languageserver": "^7.0.0",
+    "vscode-languageserver-textdocument": "1.0.2",
+    "vscode-uri": "^3.0.2"
   },
   "devDependencies": {
+    "@types/fs-extra": "9.0.13",
+    "@types/glob": "^7.2.0",
+    "@types/mocha": "^9.0.0",
+    "@types/node": "^16.11.4",
+    "@types/vscode": "^1.61.0",
     "@typescript-eslint/eslint-plugin": "^5.2.0",
     "@typescript-eslint/parser": "^5.2.0",
+    "@vscode/test-electron": "^2.4.1",
+    "@vscode/vsce": "^3.2.2",
     "eslint": "^8.1.0",
     "eslint-config-prettier": "^8.3.0",
+    "fs-extra": "^10.0.0",
+    "merge-options": "^3.0.4",
+    "mocha": "^9.1.3",
     "npm-check-updates": "^11.8.5",
     "npm-run-all": "^4.1.5",
     "prettier": "^2.4.1",
-    "sort-package-json": "^1.52.0"
+    "rimraf": "^3.0.2",
+    "sort-package-json": "^1.52.0",
+    "ts-loader": "^9.2.6",
+    "typescript": "^4.4.4",
+    "webpack": "^5.61.0",
+    "webpack-cli": "^4.9.1"
   },
   "engines": {
-    "node": ">=16.0.0"
+    "node": ">=16.0.0",
+    "vscode": "^1.74.0"
+  },
+  "icon": "textlint-icon_128x128.png",
+  "galleryBanner": {
+    "color": "#5acbe3",
+    "theme": "light"
   },
   "prettier": {
     "printWidth": 120
