--- packages/textlint-server/.eslintrc.json
@@ -1,10 +1,0 @@
-{
-  "root": true,
-  "extends": "../.eslintrc.base.json",
-  "parserOptions": {
-    "project": ["./tsconfig.json"]
-  },
-  "rules": {
-    "no-console": "error"
-  }
-}
