--- packages/textlint-server/.vscode/launch.json
@@ -1,16 +1,0 @@
-{
-  // Use IntelliSense to learn about possible Node.js debug attributes.
-  // Hover to view descriptions of existing attributes.
-  // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
-  "version": "0.2.0",
-  "configurations": [
-    {
-      "name": "Attach",
-      "type": "node",
-      "request": "attach",
-      "port": 6004,
-      "sourceMaps": true,
-      "outDir": "${workspaceRoot}/../textlint/server"
-    }
-  ]
-}
