--- packages/textlint-server/package.json
@@ -1,36 +1,0 @@
-{
-  "name": "vscode-textlint-server",
-  "version": "0.12.1",
-  "description": "Textlint Linter Server",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/textlint/vscode-textlint"
-  },
-  "license": "MIT",
-  "main": "lib/server.js",
-  "files": [
-    "lib"
-  ],
-  "scripts": {
-    "clean": "rimraf lib",
-    "tsc": "tsc -p ."
-  },
-  "dependencies": {
-    "glob": "^7.2.0",
-    "vscode-languageserver": "^7.0.0",
-    "vscode-languageserver-textdocument": "1.0.2",
-    "vscode-uri": "^3.0.2"
-  },
-  "devDependencies": {
-    "@types/glob": "^7.2.0",
-    "@types/node": "^16.11.4",
-    "rimraf": "^3.0.2",
-    "typescript": "^4.4.4"
-  },
-  "engines": {
-    "node": "*"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
