--- packages/textlint-server/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es6",
-    "module": "commonjs",
-    "moduleResolution": "node",
-    "sourceMap": true,
-    "outDir": "lib",
-    "lib": ["es6", "DOM"]
-  },
-  "exclude": ["node_modules", "lib"]
-}
