--- packages/textlint/.eslintrc.json
@@ -1,8 +1,0 @@
-{
-  "root": true,
-  "extends": "../.eslintrc.base.json",
-  "parserOptions": {
-    "project": ["./tsconfig.json"]
-  },
-  "rules": {}
-}
