--- packages/textlint/package.json
@@ -1,155 +1,0 @@
-{
-  "name": "textlint",
-  "displayName": "textlint",
-  "version": "0.12.1",
-  "description": "Integrates textlint into VS Code.",
-  "categories": [
-    "Linters"
-  ],
-  "homepage": "https://github.com/textlint/vscode-textlint",
-  "bugs": {
-    "url": "https://github.com/textlint/vscode-textlint/issues"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/textlint/vscode-textlint"
-  },
-  "license": "MIT",
-  "publisher": "3w36zj6",
-  "main": "./dist/extension",
-  "scripts": {
-    "clean": "rimraf dist out",
-    "tsc": "tsc -p ./",
-    "pretest": "npm run tsc",
-    "test": "node ./out/test/runTest.js",
-    "package": "vsce package --no-dependencies"
-  },
-  "contributes": {
-    "commands": [
-      {
-        "title": "Fix all auto-fixable Problems",
-        "category": "textlint",
-        "command": "textlint.executeAutofix"
-      },
-      {
-        "title": "Create '.textlintrc' File",
-        "category": "textlint",
-        "command": "textlint.createConfig"
-      },
-      {
-        "title": "Show Output Channel",
-        "category": "textlint",
-        "command": "textlint.showOutputChannel"
-      }
-    ],
-    "configurationDefaults": {
-      "files.associations": {
-        ".textlintrc.json": "jsonc",
-        ".textlintignore": "ignore"
-      }
-    },
-    "configuration": {
-      "type": "object",
-      "title": "textlint",
-      "properties": {
-        "textlint.languages": {
-          "default": [
-            "markdown",
-            "plaintext",
-            "html",
-            "tex",
-            "latex",
-            "doctex"
-          ],
-          "type": [
-            "array"
-          ],
-          "items": {
-            "type": "string"
-          },
-          "description": "Languages to lint with textlint."
-        },
-        "textlint.configPath": {
-          "type": "string",
-          "default": null,
-          "description": "An absolute path to textlint config file."
-        },
-        "textlint.ignorePath": {
-          "type": "string",
-          "default": null,
-          "description": "An absolute path to textlint ignore file."
-        },
-        "textlint.nodePath": {
-          "type": "string",
-          "default": null,
-          "description": "A path added to NODE_PATH when resolving the textlint module."
-        },
-        "textlint.run": {
-          "type": "string",
-          "enum": [
-            "onSave",
-            "onType"
-          ],
-          "default": "onSave",
-          "description": "Run the linter on save (onSave) or on type (onType)"
-        },
-        "textlint.autoFixOnSave": {
-          "type": "boolean",
-          "default": false,
-          "description": "Turns auto fix on save on or off."
-        },
-        "textlint.trace": {
-          "type": "string",
-          "enum": [
-            "off",
-            "messages",
-            "verbose"
-          ],
-          "default": "off",
-          "description": "Traces the communication between VSCode and the textlint linter service."
-        },
-        "textlint.targetPath": {
-          "type": "string",
-          "default": "",
-          "description": "Target files path that runs lint."
-        }
-      }
-    }
-  },
-  "activationEvents": [
-    "onLanguage",
-    "onCommand:textlint.showOutputChannel",
-    "onCommand:textlint.createConfig",
-    "onCommand:textlint.executeAutofix"
-  ],
-  "dependencies": {
-    "minimatch": "^3.0.4",
-    "vscode-languageclient": "^7.0.0",
-    "vscode-uri": "^3.0.2"
-  },
-  "devDependencies": {
-    "@types/fs-extra": "9.0.13",
-    "@types/mocha": "^9.0.0",
-    "@types/node": "^16.11.4",
-    "@types/vscode": "^1.61.0",
-    "@vscode/test-electron": "^2.4.1",
-    "@vscode/vsce": "^3.2.2",
-    "fs-extra": "^10.0.0",
-    "merge-options": "^3.0.4",
-    "mocha": "^9.1.3",
-    "npm-run-all": "^4.1.5",
-    "rimraf": "^3.0.2",
-    "ts-loader": "^9.2.6",
-    "typescript": "^4.4.4",
-    "webpack": "^5.61.0",
-    "webpack-cli": "^4.9.1"
-  },
-  "engines": {
-    "vscode": "^1.74.0"
-  },
-  "icon": "textlint-icon_128x128.png",
-  "galleryBanner": {
-    "color": "#5acbe3",
-    "theme": "light"
-  }
-}
