+++ src/test/runTest.ts
@@ -13,15 +13,13 @@ async function main() {
     // Passed to --extensionTestsPath
     const extensionTestsPath = path.resolve(__dirname, "./suite/index");
 
-    const workspaceFolder = path.resolve(extensionDevelopmentPath, "../../tests/fixtures/single-root-workspace");
+    const workspaceFolder = path.resolve(extensionDevelopmentPath, "tests/fixtures/single-root-workspace");
 
     try {
-      console.log("Running npm ci...");
       execSync("npm ci", {
         cwd: workspaceFolder,
         stdio: "inherit",
       });
-      console.log("Dependencies installed successfully");
     } catch (error) {
       console.error("Failed to install dependencies with npm ci");
       console.error(error);
