+++ tsconfig.json
@@ -6,5 +6,6 @@
     "lib": ["es6", "DOM"],
     "sourceMap": true
   },
-  "exclude": ["node_modules", ".vscode-test"]
+  "include": ["src/**/*.ts"],
+  "exclude": ["node_modules", ".vscode-test", "dist", "out"]
 }
